+++ src/main/java/gift/controller/ProductController.java
@@ -1,9 +1,9 @@
-package gift;
+package gift.controller;
 
+import gift.model.Product;
+import gift.service.ProductService;
 import jakarta.validation.Valid;
-import jakarta.validation.ValidationException;
 import java.util.List;
-import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
