+++ src/main/resources/templates/Edit_product.html
@@ -6,24 +6,21 @@
 </head>
 <body>
 <h1>상품 수정</h1>
-<form th:action="@{/admin/products}" th:object="${product}" method="post">
+
+<form th:action="@{/admin/products/{id}(id=${product.id})}" th:object="${product}" method="post">
+  <input type="hidden" name="_method" value="PUT" /> <!-- PUT 요청을 처리하기 위한 숨겨진 필드 -->
   <div>
-    <label for="id">Id:</label>
-    <input type="number" id="id" th:field="*{id}" required />
-    <div th:if="${#fields.hasErrors('id')}" th:errors="*{id}" class="error-message"></div>
-  </div>
-  <div>
-    <label for="name">Name:</label>
+    <label for="name">이름:</label>
     <input type="text" id="name" th:field="*{name}" required />
     <div th:if="${#fields.hasErrors('name')}" th:errors="*{name}" class="error-message"></div>
   </div>
   <div>
-    <label for="price">Price:</label>
+    <label for="price">가격:</label>
     <input type="number" id="price" th:field="*{price}" required />
     <div th:if="${#fields.hasErrors('price')}" th:errors="*{price}" class="error-message"></div>
   </div>
   <div>
-    <label for="imageUrl">Image URL:</label>
+    <label for="imageUrl">이미지 URL:</label>
     <input type="text" id="imageUrl" th:field="*{imageUrl}" required />
     <div th:if="${#fields.hasErrors('imageUrl')}" th:errors="*{imageUrl}" class="error-message"></div>
   </div>
